// src/app/modules/financeiro/banco/enums/cadastros-banco-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 26.0.4
Code generated at time stamp: 2019-10-26T12:47:23.552
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface SituacaoConciliacaoTrn extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface SituacaoConciliacao extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoTransacao extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoContaBancaria extends SelectItem {
	
  label?: string;
  value: any;
  
}